Rename address discovery action creator
